refactor: mybatis 升级为 mybatis-plus refactor: 表结构拆分 feat: druid增加stat 统计慢sql feat: mybatis增加完整慢sql打印 feat: 使用logback完善日志 feat: 增加userId traceId token的传递 test: 完善单元测试

+++ api/src/main/java/com/gill/api/domain/UserProperties.java
@@ -87,11 +87,11 @@ public class UserProperties {
         return REDIS_TOKEN_PREFIX + token;
     }
 
-    public static String getRedisUserInfoKey(int userId) {
+    public static String getRedisUserInfoKey(long userId) {
         return REDIS_USER_INFO_PREFIX + userId;
     }
 
-    public static String getRedisUserResourceKey(int userId) {
+    public static String getRedisUserResourceKey(long userId) {
         return REDIS_USER_RESOURCE_PREFIX + userId;
     }
 }
